refactor: Rename GoToTop component and update import paths
